+++ CHANGELOG.rst
@@ -30,6 +30,7 @@ Fixed
     - Alert email not sent for users without primary address (#1607)
 - **Projectroles**
     - ``check_app_setting_defs()`` crash with no database connection (#1615)
+    - Sporadic UI test failures in ``test_submit_button()`` (#1618)
 
 
 v1.1.1 (2025-03-11)